Rename pkg/rpcapp to pkg/app
Shorter, cleaner package name. rpcApp struct -> App.
