+++ CHANGELOG.md
@@ -5,13 +5,13 @@ Not a git log mirror — focus on what changed at the design level and why.
 
 ## Architecture: Code Organization Refactor
 
-### cmd/ai → pkg/rpcapp + pkg/cli (2026-05)
+### cmd/ai → pkg/app + pkg/cli (2026-05)
 
 **Problem**: All RPC handler logic lived in `package main` (cmd/ai), making it untestable.
 cmd/ai had grown to 5700+ lines across 20+ files.
 
 **Changes**:
-- Moved all `rpc_*.go`, `session_writer.go`, `helpers.go` to `pkg/rpcapp/`
+- Moved all `rpc_*.go`, `session_writer.go`, `helpers.go` to `pkg/app/`
 - Moved CLI subcommands (run/serve/ls/send/kill/watch) to `pkg/cli/`
 - `cmd/ai/main.go` is now a thin 182-line entry point
 - Added smoke tests that exercise the full RunRPC pipeline (coverage 6.8% → 44.5%)
