+++ pkg/app/handler_steer_followup_test.go
@@ -1,4 +1,4 @@
-package rpcapp
+package app
 
 import (
 	"encoding/json"
