+++ pkg/app/rpc_app.go
@@ -1,4 +1,4 @@
-package rpcapp
+package app
 
 import (
 	"encoding/json"
