+++ pkg/app/rpc_config_handlers.go
@@ -1,4 +1,4 @@
-package rpcapp
+package app
 
 import (
 	"fmt"
