+++ pkg/app/rpc_help_handlers.go
@@ -1,4 +1,4 @@
-package rpcapp
+package app
 
 import (
 	"fmt"
