+++ pkg/app/rpc_message_handlers.go
@@ -1,4 +1,4 @@
-package rpcapp
+package app
 
 import (
 	"context"
