+++ pkg/app/rpc_session_handlers_test.go
@@ -1,4 +1,4 @@
-package rpcapp
+package app
 
 import (
 	"testing"
