+++ pkg/app/rpc_setup.go
@@ -1,4 +1,4 @@
-package rpcapp
+package app
 
 import (
 	"fmt"
