+++ pkg/app/rpcapp_smoke_test.go
@@ -1,4 +1,4 @@
-package rpcapp
+package app
 
 import (
 	"bufio"
