+++ pkg/app/session_writer.go
@@ -1,4 +1,4 @@
-package rpcapp
+package app
 
 import (
 	"context"
